backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic==2.10.4
numpy==2.2.1
pillow==11.1.0
torch==2.11.0
torchvision==0.26.0
scikit-learn==1.6.0
opencv-python==4.10.0.84
shap==0.46.0
lime==0.2.0.1
